Added many of the things needed to NormodFile.  Now can print the particle lines soon the hole lines.
Needs work to distinguish when printing emptiness and when printing fullness.  Should change the methods in
Shell to return the fullness and emptiness instead of occupancy.  Think I just need to sqaure it.

Added some methods to transition.  Now can return a vector of the reaction without duplicates.  So if it is (p,n)
GetFirstShells would return the neutron shells that are participating with out duplicates.